@@ -6,29 +6,42 @@
     "dev": "next dev",
     "build": "next build",
     "start": "next start",
-    "lint": "next lint"
+    "deploy": "next build && firebase deploy --only functions,hosting"
   },
   "dependencies": {
     "@react-google-maps/api": "^2.19.2",
     "@vercel/analytics": "^1.3.1",
     "@vercel/speed-insights": "^1.0.12",
-    "firebase": "^9.23.0",
-    "next": "13.5.1",
-    "openai": "^4.0.0",
-    "react": "18.2.0",
-    "react-dom": "18.2.0",
-    "react-markdown": "^8.0.7",
-    "twilio": "^4.18.0",
     "autoprefixer": "10.4.15",
     "eslint": "8.49.0",
     "eslint-config-next": "13.5.1",
+    "firebase": "^9.23.0",
+    "firebase-admin": "^12.6.0",
+    "firebase-functions": "^6.0.1",
+    "next": "^14.2.15",
+    "openai": "^4.0.0",
     "postcss": "8.4.30",
+    "react": "^18.3.1",
+    "react-dom": "^18.3.1",
+    "react-icons": "^5.3.0",
+    "react-markdown": "^8.0.7",
     "tailwindcss": "3.3.3",
+    "twilio": "^4.18.0",
     "typescript": "5.2.2"
   },
   "devDependencies": {
     "@types/node": "20.6.2",
     "@types/react": "18.2.22",
-    "@types/react-dom": "18.2.7"
-  }
+    "@types/react-dom": "18.2.7",
+    "firebase-functions-test": "^3.3.0"
+  },
+  "description": "This project is set up for automatic deployment with Vercel.",
+  "main": "deploy.js",
+  "directories": {
+    "doc": "docs",
+    "lib": "lib"
+  },
+  "keywords": [],
+  "author": "",
+  "license": "ISC"
 }
