@@ -1,9 +1,8 @@
 /** @type {import('next').NextConfig} */
 const nextConfig = {
   reactStrictMode: true,
-  swcMinify: false, // Disable SWC minification
   images: {
-    domains: ['placeholder.com']
+    domains: ['firebasestorage.googleapis.com'],
   },
 };
 
